Added LB302 BassSynth, 302 sample project, and modified Knob for 64bit systems. Read plugins/lb302/README for more information.
git-svn-id: https://lmms.svn.sf.net/svnroot/lmms/trunk/lmms@455 0778d3d1-df1d-0410-868b-ea421aaaa00d

+++ src/widgets/knob.cpp
@@ -469,8 +469,10 @@ void knob::mousePressEvent( QMouseEvent * _me )
 /*			eng()->getMainWindow()->isCtrlPressed() == TRUE &&*/
 			eng()->getMainWindow()->isShiftPressed() == TRUE )
 	{
+        /* this pointer was casted to uint, 
+         * compile time error on 64 bit systems */
 		new stringPairDrag( "link_object",
-						QString::number( (uint) this ),
+						QString::number( (ulong) this ),
 						QPixmap(), this, eng() );
 	}
 	else if( _me->button() == Qt::MidButton )
